requirements.txt
alembic
anyio
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apispec
argcomplete
argon2-cffi
argon2-cffi-bindings
arrow
attrs
Babel
bcrypt
blinker
Brotli
cachelib
cassandra-driver==3.26.0
cattrs
certifi
cffi
charset-normalizer
click==8.1.3
clickclick
colorama
colorlog==4.8.0
colour==0.1.5
ConfigUpdater
connexion
contourpy==1.0.7
cqlsh==6.1.2
cron-descriptor
croniter
cryptography
cycler==0.11.0
Deprecated
dill
dnspython
docutils
email-validator
exceptiongroup
findspark==2.0.1
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
fonttools==4.39.4
furl
gate-api==4.60.2
geomet==0.2.1.post1
graphviz
greenlet
gunicorn
h11
h2
hpack==4.0.0
httpcore
httpx
hyperframe==6.0.1
idna
importlib-metadata
infinity==1.5
inflection==0.5.1
intervals==0.9.2
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
kafka-python==2.0.2
kiwisolver==1.4.4
lazy-object-proxy
linkify-it-py
lockfile==0.12.2
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
marshmallow-oneofschema
marshmallow-sqlalchemy
matplotlib==3.7.1
mdit-py-plugins
mdurl
numpy==1.24.3
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.1
passlib==1.7.4
pathspec
pendulum
phonenumbers
Pillow==9.5.0
pluggy
prison
psutil
py4j==0.10.9.7
pycparser
Pygments
PyJWT
pyOpenSSL
PySocks
pyspark==3.4.1
python-daemon
python-dateutil==2.8.2
python-nvd3
python-slugify
pytz==2023.3
pytzdata
PyYAML
referencing
requests
requests-toolbelt
rfc3986
rich
rpds-py
setproctitle
six
sniffio
spark==0.2.1
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
swagger-ui-bundle
tabulate
tenacity
termcolor
text-unidecode
typing_extensions
tzdata==2023.3
uc-micro-py
unicodecsv==0.14.1
Unidecode
urllib3
Werkzeug
wrapt
WTForms
zipp
